fixed debug errors

+++ assets/camera_controller.gd
@@ -35,6 +35,8 @@ func _on_set_movestate(movestate: MoveState):
 	move_tween.tween_property(camera, "fov", movestate.fov, .5).set_trans(Tween.TRANS_SINE).set_ease(Tween.EASE_OUT)
 	
 func _on_set_playerstate(playerstate: PlayerStateChange):
+	if playerstate.new_state.id == playerstate.old_state.id:
+		return
 	if player_tween:
 		player_tween.kill()
 	player_tween = create_tween()
